Rename: Tooltip 스타일 전역으로 분리
